requirements.txt
streamlit>=1.40.0
plotly>=5.0.0
pandas>=2.0.0
openai>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
openpyxl>=3.1.0
python-docx>=1.1.0
